Remove duplication from method predicates
Issue: #14
